requirements.txt
pillow
opencv-python
scikit-image
numpy
scipy
matplotlib
seaborn
python-dotenv
tqdm
imageio
streamlit
google-genai